+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - molecules-search-search_field: added search field form component.
+- atoms-checkbox: added checkbox form component.
 
 ### Removed